Add consent factory test + refactor
