+++ src/cmp/domain/ConsentFactory.js
@@ -1,12 +1,8 @@
-import {ConsentString} from 'consent-string'
-
+/**
+ * @interface
+ */
 export default class ConsentFactory {
   createConsent({consentString, globalVendorList}) {
-    return Promise.resolve()
-      .then(() => new ConsentString(consentString))
-      .then(consent => {
-        consent.setGlobalVendorList(globalVendorList)
-        return consent
-      })
+    throw new Error('ConsentFactory#createConsent must be implemented')
   }
 }
