+++ test/cmp/infrastructure/factory/IABConsentFactoryTest.js
@@ -1,0 +1,22 @@
+/* eslint-disable prettier/prettier */
+import {expect} from 'chai'
+import IABConsentFactory from "../../../../src/cmp/infrastructure/factory/IABConsentFactory";
+
+describe('IAB Consent Factory', () => {
+    describe('createConsent method', () => {
+        it('Should return a ConsentString with a globalVendorList', done => {
+            const givenGlobalVendorList = require('./../../resources/globalvendorlist.json')
+            const givenConsentStringData = 'BOPVloMOPi60FABABAENBA-AAAAcF7_______9______9uz_Gv_r_f__33e8_39v_h_7_-___m_-3zV4-_lvR11yPA1OrfIrwFhiAwAA'
+            const factory = new IABConsentFactory()
+            factory.createConsent({
+                consentString: givenConsentStringData,
+                globalVendorList: givenGlobalVendorList
+            }).then(consent => {
+                expect(consent.getVendorListVersion(), 'consent has no valid vendor list version').to.equal(67)
+                expect(consent.getCmpId(), 'consent has not parsed the string data').to.equal(1)
+            })
+            .then(() => done())
+            .catch(e => done(e))
+        })
+    })
+})
